+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 ### Bug Fixes
 Solved code smells detected by Sonar ([1a704ba](https://github.com/OntimizeWeb/ontimize-web-ngx-gallery/commit/1a704ba)) Closes [#34](https://github.com/OntimizeWeb/ontimize-web-ngx-gallery/issues/34)
 ### BREAKING CHANGES
-***o-gallery-options**: now `no` values ​​of `imageSwipe`, `thumbnailsSwipe`, `previewSwipe` values `no` have been replaced by `yes` ([5c447fa](https://github.com/OntimizeWeb/ontimize-web-ngx-gallery/commit/5c447fa)) Closes [#34](https://github.com/OntimizeWeb/ontimize-web-ngx-gallery/issues/34)
+***o-gallery-options**: now `no` values ​​of `imageSwipe`, `thumbnailsSwipe`, `previewSwipe` values `no` have been replaced by `yes` ([5c447fa](https://github.com/OntimizeWeb/ontimize-web-ngx-gallery/commit/5c447fa)) Closes [#36](https://github.com/OntimizeWeb/ontimize-web-ngx-gallery/issues/36)
 
 ## 8.0.3 (2022-07-07)
 ### Bug Fixes
-* **GalleryImage**: fix image type detection when url contains query parameters ([983376d](https://github.com/OntimizeWeb/ontimize-web-ngx-gallery/commit/983376d2b08154fd67eba224289f4f931ad17246)) Closes [#36](https://github.com/OntimizeWeb/ontimize-web-ngx-gallery/issues/36)
+* **GalleryImage**: fix image type detection when url contains query parameters ([983376d](https://github.com/OntimizeWeb/ontimize-web-ngx-gallery/commit/983376d2b08154fd67eba224289f4f931ad17246)) Closes [#42](https://github.com/OntimizeWeb/ontimize-web-ngx-gallery/issues/42)
 
 ## 8.0.2 (2021-10-27)
 ### Features