+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 ## 8.1.0
+### Features
+* **o-gallery**: New methods `changeImageSwipe`, `changeThumbnailsSwipe`,`changePreviewSwipe` ([92ca941](https://github.com/OntimizeWeb/ontimize-web-ngx-gallery/commit/92ca941))
 ### Bug Fixes
 Solved code smells detected by Sonar ([1a704ba](https://github.com/OntimizeWeb/ontimize-web-ngx-gallery/commit/1a704ba)) Closes [#34](https://github.com/OntimizeWeb/ontimize-web-ngx-gallery/issues/34)
 ### BREAKING CHANGES
-***o-gallery-options**: now `no` values ​​of `imageSwipe`, `thumbnailsSwipe`, `previewSwipe` values `no` have been replaced by `yes` ([5c447fa](https://github.com/OntimizeWeb/ontimize-web-ngx-gallery/commit/5c447fa)) Closes [#36](https://github.com/OntimizeWeb/ontimize-web-ngx-gallery/issues/36)
+* **o-gallery-options**:
+  * now `no` values ​​of `imageSwipe`, `thumbnailsSwipe`, `previewSwipe` values `no` have been replaced by `yes` ([5c447fa](https://github.com/OntimizeWeb/ontimize-web-ngx-gallery/commit/5c447fa)) Closes [#36](https://github.com/OntimizeWeb/ontimize-web-ngx-gallery/issues/36)
+  * now it is not necessary to add the properties with the default values ([ffea97c](https://github.com/OntimizeWeb/ontimize-web-ngx-gallery/commit/ffea97c))
+
 
 ## 8.0.3 (2022-07-07)
 ### Bug Fixes